\name{csampling-package}
\alias{csampling-package}
\alias{csampling}
\docType{package}
\title{
Functions for Conditional Simulation in Regression-Scale Models
}
\description{
Monte Carlo conditional inference for the parameters of a
linear nonnormal regression model
}
\details{
\tabular{ll}{
Package: \tab csampling\cr
Version: \tab 1.0-0\cr
Date: \tab 2005-05-31\cr
Depends: \tab R (>= 2.0.1), statmod, survival\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
License: \tab GPL (version 2 or later)\cr
URL: \tab http://www.r-project.org, http://www.isib.cnr.it/~brazzale/CS/\cr
Bundle: \tab hoa\cr
BundleDescription: \tab Various functions for higher order likelihood-based
inference\cr
Contains: \tab cond marg nlreg csampling\cr
Packaged: \tab Thu Nov 24 11:57:28 2005; brazzale\cr
Built: \tab R 2.2.0; ; 2005-11-24 12:36:48; unix\cr
}

Index:
\preformatted{
Functions:
=========
Laplace                 Calculate Laplace's Marginal Density
                        Approximation
dmt                     Multivariate Student t Distribution
make.sample.data        Create a Conditional Sampling Data Object
plot.Lapl.spl           Plot uni- and bivariate approximate marginal
                        densities
rsm.sample              Conditional Sampler for Regression-Scale
                        Models
}
}
\author{
S original by Alessandra R. Brazzale <alessandra.brazzale@unimore.it>.  
R port by Alessandra R. Brazzale <alessandra.brazzale@unimore.it>.

Maintainer: Alessandra R. Brazzale <alessandra.brazzale@unimore.it>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
