\name{plot.Lapl}
\alias{plot.Lapl.spl}
\alias{plot.Lapl.cont}
\title{
  Plot uni- and bivariate approximate marginal densities
}
\description{
  Plots the uni- and bivariate approximations to the marginal 
  densities of components of the \acronym{MLE} obtained by Laplace's 
  method.
}
\usage{
\method{plot}{Lapl.spl}(x, \dots)
\method{plot}{Lapl.cont}(x, \dots)
}
\arguments{
  \item{x}{
    an object of class \code{Lapl.spl} or \code{Lapl.cont} such as
    generated by the \code{\link[csampling]{Laplace}} function.
  }
  \item{\dots}{
    additional graphics parameters.
  }
}
\details{
  This is a method for the function \code{plot()} for objects 
  inheriting from class \code{Lapl.spl} and \code{Lapl.cont} 
  generated by the \code{Laplace()} routine.  
}
\seealso{
  \code{\link[csampling]{Laplace}}
}
\keyword{hplot}
\keyword{methods}

