\name{csampling-package}
\Rdversion{1.1}
\alias{csampling-package}
\alias{csampling}
\docType{package}
\title{
Functions for Conditional Simulation in Regression-Scale Models
}
\description{
Monte Carlo conditional inference for the parameters of a
linear nonnormal regression model
}
\details{
\tabular{ll}{
Package: \tab csampling\cr
Version: \tab 1.2-0\cr
Date: \tab 2009-10-03\cr
Depends: \tab R (>= 2.6.0), marg, statmod, survival\cr
License: \tab GPL (>= 2)\cr
URL: \tab http://www.r-project.org, http://statwww.epfl.ch/AA/\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}

Index:
\preformatted{
Functions:
=========
Laplace                 Calculate Laplace's Marginal Density
                        Approximation
dmt                     Multivariate Student t Distribution
make.sample.data        Create a Conditional Sampling Data Object
plot.Lapl.spl           Plot uni- and bivariate approximate marginal
                        densities
rsm.sample              Conditional Sampler for Regression-Scale
                        Models
}
}
\author{
S original by Alessandra R. Brazzale <alessandra.brazzale@unimore.it>.
R port by Alessandra R. Brazzale <alessandra.brazzale@unimore.it>.

Maintainer: Alessandra R. Brazzale <alessandra.brazzale@unimore.it>
}
\keyword{ package }
